libwasi: move environ/args functions to a separate file
